Fix namespaces
