+++ src/Adapter/PDF/PDFGenerator.php
@@ -27,7 +27,7 @@
 namespace PrestaShop\PrestaShop\Adapter\PDF;
 
 use PDF;
-use PrestaShop\PrestaShop\Core\DataProvider\PDFTemplateTypeProviderInterface;
+use PrestaShop\PrestaShop\Core\PDF\PDFTemplateTypeProviderInterface;
 use PrestaShop\PrestaShop\Core\PDF\PDFGeneratorInterface;
 use Smarty;
