+++ src/PrestaShopBundle/Form/Admin/Sell/Order/Invoices/InvoiceByDateFormHandler.php
@@ -27,7 +27,7 @@
 namespace PrestaShopBundle\Form\Admin\Sell\Order\Invoices;
 
 use DateTime;
-use PrestaShop\PrestaShop\Core\DataProvider\OrderInvoiceDataProviderInterface;
+use PrestaShop\PrestaShop\Core\Order\OrderInvoiceDataProviderInterface;
 use PrestaShop\PrestaShop\Core\Form\FormDataProviderInterface;
 use PrestaShop\PrestaShop\Core\Form\FormHandler;
 use PrestaShop\PrestaShop\Core\PDF\PDFGeneratorInterface;
